fix(model_profiles): break the reviewer_value import cycle, add test mirrors
Three hard convention violations from iteration 1.

no_circular_imports: reviewer_value imported _fold_resolved_model from
model_profiles_storage, which lazily imports reviewer_value's folds from
apply_run. The function is stdlib-only and writes the by_resolved_model
breakdown whose key constants already live in model_profiles_identity, so it
moves there — one shape, one writer, below every accumulator that writes it
rather than inside one of them. Three accumulators in two modules write that
breakdown; a breakdown written three ways stops summing to the denominator it
explains. Storage and reviewer_value now both reach it through identity, and
the facade re-exports it from its new owner.

test_mirrors_source: adds tests/test_model_profiles_identity.py (the shared
vocabulary, exercised against plain dicts) and tests/test_model_profiles_storage.py
(the accumulation tests, moved out of test_model_profiles.py, which is left as
the facade compatibility + seam file and thereby also drops back under
max_test_file_lines).

test_model_profiles_boundary.py gains TestNoCycleWithReviewerValue so the cycle
cannot return. Moved code remains byte-identical to main (AST-verified).
